central-node/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
